Melakukan refactoring Long Parameter List
Melakukan refactoring Long Parameter List dengan menginisialisasi variable dan merubah penulisan parameter dari class employee ke class DataPribadi, GajiPegawai, KeluargaPegawai, TahunPegawai
