+++ TaxCalculator/src/lib/TahunPegawai.java
@@ -1,0 +1,14 @@
+package lib;
+
+public class TahunPegawai {
+    private int yearJoined;
+	private int monthJoined;
+	private int dayJoined;
+	private int monthWorkingInYear;
+
+    public TahunPegawai(int yearJoined, int monthJoined, int dayJoined, int monthWorkingInYear) {
+		this.monthJoined = monthJoined;
+		this.dayJoined = dayJoined;
+        this.monthWorkingInYear = monthWorkingInYear;
+	}
+}
